+++ pytsite/core/assetman/__init__.py
@@ -24,4 +24,5 @@ __init()
 
 
 # Public API
-from ._functions import register_package, add, add_location, remove, remove_location, dump_js, dump_css
+from ._functions import register_package, add, add_location, remove, remove_location, dump_js, dump_css,\
+    get_url