+++ pytsite/page/_model.py
@@ -1,0 +1,11 @@
+"""Page Models.
+"""
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
+from pytsite import content as _content
+
+
+class Page(_content.model.Content):
+    pass