+++ CHANGELOG.md
@@ -8,8 +8,11 @@
 - `args` argument in `core.odm.Model.t()`
 - Events in ODM model: `odm.entity.pre_save`, `odm.entity.save` 
 - Events in Content ODM model: `content.entity.pre_save`, `content.entity.save`
-- Properties in `content.model.Content`: `title`, `description`, `body`, `url`, `tags`, `images`, `links`
-- Properties in `taxonomy.model.Term`: `title`, `alias`, `language`, `weight`, `order` 
+- Properties in `content.model.Content`: `title`, `description`, `body`, `url`, `tags`, `images`, `links`,
+  `author`, `section`.
+- Properties in `taxonomy.model.Term`: `title`, `alias`, `language`, `weight`, `order`.
+- Properties in `auth.model.User`: `full_name`.
+- Properties in `file.model.File`: `url`.
 - `stage` argument in `pytsite.odm_ui.UIMixin.setup_m_form()` hook.
 - `args` argument in `pytsite.core.odm.Model.t()`.
 - Now possible to specify package name without 'pytsite.' prefix in `core.lang.t()`, `core.router.endpoint_url`,
@@ -19,6 +22,8 @@
 - `autodetect` argument in `geo.widget.SearchAddress()`.
 - `core.validation.rule.ListListItemNotEmpty()`, `core.validation.rule.ListListItemUrl()`,
   `core.validation.rule.VideoHostingUrl()`.
+- Template engine globals: `url`, `current_url`, `base_url` , `endpoint_url`.
+  
 
 ### Changed
 - Config parameter changed `auth.auto_signup` -> `auth.allow_signup`